Implement changes suggested in PR comments
